requirements.txt
djangorestframework
humanfriendly
python-magic
